requirements.txt
alembic==1.8.1
aniso8601==9.0.1
attrs==22.1.0
bcrypt==4.0.1
bidict==0.22.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
defusedxml==0.7.1
dnspython==1.16.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Migrate==4.0.0
flask-restx==1.0.3
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.1
httpie==3.2.1
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
mariadb==1.1.5.post3
MarkupSafe==2.1.1
multidict==6.0.2
packaging==21.3
pkgutil_resolve_name==1.3.10
protobuf==3.20.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.6
requests==2.28.1
requests-toolbelt==0.10.1
rich==12.6.0
six==1.16.0
SQLAlchemy==1.4.44
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
wincertstore==0.2
WTForms==3.0.1
zipp==3.11.0
zope.event==4.5.0
zope.interface==5.5.2
